Cambio de contraseña y verificacion de correo

+++ app/Http/Controllers/Api/Auth/LoginController.php
@@ -41,6 +41,15 @@ class LoginController extends Controller
             $user=DB::table('users as u')->where('u.email',request('username'))
             ->first();
             if($user != null){
+
+                //validar si el usuario esta verificado y a cambiado su contraseña
+                if($user->estado == 0){
+                    return response()->json([
+                        'message' => 'Por favor actualizar su contraseña',
+                        'url' => 'passwordUpdate'
+                    ], 200);
+                }
+
                 if (Hash::check(request('password'), $user->password)) {
                     DB::table('oauth_access_tokens')->where('user_id', $user->id)->delete();
